requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
openai==1.12.0
python-dotenv==1.0.1
openai-whisper==20231117
python-multipart==0.0.9
ollama==0.1.6
streamlink==6.6.1
ffmpeg-python==0.2.0
faster-whisper==0.10.0
aiofiles==23.2.1
twitchio==2.8.2
pydantic-settings==2.1.0
streamlink==6.6.1
